# core-lending-collateral-management-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-lending-collateral-management
  description: Collateral & Guarantee Management Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/general-ledger
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
  /api/v1/guarantees/{guaranteeRecordId}:
    get:
      tags:
        - GuaranteeRecord
      summary: Get a GuaranteeRecord by ID
      operationId: getById
      parameters:
        - name: guaranteeRecordId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/GuaranteeRecordDTO'
    put:
      tags:
        - GuaranteeRecord
      summary: Update a GuaranteeRecord
      operationId: update
      parameters:
        - name: guaranteeRecordId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuaranteeRecordDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/GuaranteeRecordDTO'
    delete:
      tags:
        - GuaranteeRecord
      summary: Delete a GuaranteeRecord
      operationId: delete
      parameters:
        - name: guaranteeRecordId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
  /api/v1/guarantees/{guaranteeRecordId}/valuations/{guaranteeValuationId}:
    get:
      tags:
        - GuaranteeValuation
      summary: Get a Guarantee Valuation by ID
      operationId: getById_1
      parameters:
        - name: guaranteeRecordId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: guaranteeValuationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/GuaranteeValuationDTO'
    put:
      tags:
        - GuaranteeValuation
      summary: Update a Guarantee Valuation
      operationId: update_1
      parameters:
        - name: guaranteeRecordId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: guaranteeValuationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuaranteeValuationDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/GuaranteeValuationDTO'
    delete:
      tags:
        - GuaranteeValuation
      summary: Delete a Guarantee Valuation
      operationId: delete_1
      parameters:
        - name: guaranteeRecordId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: guaranteeValuationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
...
  /api/v1/collateral-cases/{collateralCaseId}/assets/{collateralAssetId}/valuations/{collateralValuationId}:
    get:
      tags:
        - CollateralValuation
      summary: Get a Collateral Valuation by ID
      operationId: getById_8
      parameters:
        - name: collateralCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: collateralAssetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: collateralValuationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CollateralValuationDTO'
    put:
      tags:
        - CollateralValuation
      summary: Update a Collateral Valuation
      operationId: update_8
      parameters:
        - name: collateralCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: collateralAssetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: collateralValuationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollateralValuationDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CollateralValuationDTO'
    delete:
      tags:
        - CollateralValuation
      summary: Delete a Collateral Valuation
      operationId: delete_8
      parameters:
        - name: collateralCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: collateralAssetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: collateralValuationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
...
  /api/v1/collateral-assets/{collateralAssetId}/property-details/{propertyId}:
    get:
      tags:
        - RealEstateProperty
      summary: Get Real Estate Property Details by ID
      operationId: getById_13
      parameters:
        - name: collateralAssetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RealEstatePropertyDTO'
    put:
      tags:
        - RealEstateProperty
      summary: Update Real Estate Property Details
      operationId: update_13
      parameters:
        - name: collateralAssetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RealEstatePropertyDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RealEstatePropertyDTO'
    delete:
      tags:
        - RealEstateProperty
      summary: Delete Real Estate Property Details
      operationId: delete_13
      parameters:
        - name: collateralAssetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
...
  /api/v1/guarantees/{guaranteeRecordId}/valuations:
    get:
      tags:
        - GuaranteeValuation
      summary: List or Search Guarantee Valuations
      operationId: findAll_1
      parameters:
        - name: guaranteeRecordId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: filterRequest
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/FilterRequestGuaranteeValuationDTO'
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseGuaranteeValuationDTO'
    post:
      tags:
        - GuaranteeValuation
      summary: Create a Guarantee Valuation
      operationId: create_1
      parameters:
        - name: guaranteeRecordId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuaranteeValuationDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/GuaranteeValuationDTO'
...
  /api/v1/collateral-assets/{collateralAssetId}/property-details:
    get:
      tags:
        - RealEstateProperty
      summary: List or Search Real Estate Property Details for an Asset
      operationId: findAll_13
      parameters:
        - name: collateralAssetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: filterRequest
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/FilterRequestRealEstatePropertyDTO'
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseRealEstatePropertyDTO'
    post:
      tags:
        - RealEstateProperty
      summary: Create Real Estate Property Details under the given Asset
      operationId: create_13
      parameters:
        - name: collateralAssetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RealEstatePropertyDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RealEstatePropertyDTO'
...
components:
  schemas:
    GuaranteeRecordDTO:
      required:
        - guaranteeAmount
        - guaranteeStatus
        - guaranteeType
        - loanApplicationId
        - startDate
        - termsConditions
      type: object
      properties:
        guaranteeRecordId:
          type: string
          format: uuid
          readOnly: true
        loanApplicationId:
          type: string
          format: uuid
        loanServicingCaseId:
          type: string
          format: uuid
        guaranteeType:
          type: string
          enum:
            - PERSONAL
            - CORPORATE
            - STANDBY_LETTER
            - BANK_GUARANTEE
        guaranteeAmount:
          minimum: 0
          exclusiveMinimum: true
          type: number
        guaranteeStatus:
          type: string
          enum:
            - ACTIVE
            - EXPIRED
            - TERMINATED
            - PENDING
        termsConditions:
          maxLength: 2000
          minLength: 1
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    GuaranteeValuationDTO:
      required:
        - assessedRiskAmount
        - guaranteeRecordId
        - rationale
        - riskGrade
      type: object
      properties:
        guaranteeValuationId:
          type: string
          format: uuid
          readOnly: true
        guaranteeRecordId:
          type: string
          format: uuid
        assessedRiskAmount:
          minimum: 0
          exclusiveMinimum: false
          type: number
        riskGrade:
          type: string
          enum:
            - LOW
            - MEDIUM
            - HIGH
        rationale:
          maxLength: 2000
          minLength: 1
          type: string
        valuationDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CollateralValuationDTO:
      required:
        - appraisedValue
        - collateralAssetId
        - currencyCode
        - valuationDate
        - valuationMethod
        - valuationProvider
      type: object
      properties:
        collateralValuationId:
          type: string
          format: uuid
          readOnly: true
        collateralAssetId:
          type: string
          format: uuid
        appraisedValue:
          minimum: 0
          exclusiveMinimum: true
          type: number
        valuationMethod:
          type: string
          enum:
            - MARKET_COMPARISON
            - INCOME_APPROACH
            - COST_APPROACH
            - OTHER
        valuationProvider:
          maxLength: 200
          minLength: 1
          type: string
        appraiserName:
          maxLength: 255
          minLength: 0
          type: string
        appraiserLicense:
          maxLength: 100
          minLength: 0
          type: string
        valuationDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        currencyCode:
          type: string
          enum:
            - EUR
            - USD
            - GBP
        marketValue:
          minimum: 0
          exclusiveMinimum: false
          type: number
        rentalValue:
          minimum: 0
          exclusiveMinimum: false
          type: number
        replacementCost:
          minimum: 0
          exclusiveMinimum: false
          type: number
        landValue:
          minimum: 0
          exclusiveMinimum: false
          type: number
        buildingValue:
          minimum: 0
          exclusiveMinimum: false
          type: number
        salvageValue:
          minimum: 0
          exclusiveMinimum: false
          type: number
        comparableAssets:
          type: string
        methodology:
          maxLength: 2000
          minLength: 0
          type: string
        assumptions:
          maxLength: 2000
          minLength: 0
          type: string
        limitations:
          maxLength: 2000
          minLength: 0
          type: string
        requiresRepairs:
          type: boolean
        requiredRepairs:
          maxLength: 2000
          minLength: 0
          type: string
        repairCost:
          minimum: 0
          exclusiveMinimum: false
          type: number
        notes:
          maxLength: 1000
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    RealEstatePropertyDTO:
      required:
        - addressLine1
        - city
        - collateralAssetId
        - countryCode
        - propertyType
      type: object
      properties:
        propertyId:
          type: string
          format: uuid
          readOnly: true
        collateralAssetId:
          type: string
          format: uuid
        propertyType:
          type: string
          enum:
            - RESIDENTIAL
            - COMMERCIAL
            - LAND
            - INDUSTRIAL
            - MIXED_USE
            - AGRICULTURAL
        propertyUse:
          type: string
          enum:
            - PRIMARY_RESIDENCE
            - SECONDARY_RESIDENCE
            - INVESTMENT
            - COMMERCIAL_USE
        propertyCondition:
          type: string
          enum:
            - EXCELLENT
            - GOOD
            - FAIR
            - POOR
            - UNDER_CONSTRUCTION
        addressLine1:
          maxLength: 255
          minLength: 1
          type: string
        addressLine2:
          maxLength: 255
          minLength: 0
          type: string
        city:
          maxLength: 100
          minLength: 1
          type: string
        state:
          maxLength: 100
          minLength: 0
          type: string
        postalCode:
          maxLength: 20
          minLength: 0
          type: string
        countryCode:
          maxLength: 3
          minLength: 2
          type: string
        landArea:
          minimum: 0
          exclusiveMinimum: true
          type: number
        landAreaUnit:
          maxLength: 20
          minLength: 0
          type: string
        builtArea:
          minimum: 0
          exclusiveMinimum: true
          type: number
        builtAreaUnit:
          maxLength: 20
          minLength: 0
          type: string
        constructionYear:
          type: integer
          format: int32
        renovationYear:
          type: integer
          format: int32
        titleNumber:
          maxLength: 100
          minLength: 0
          type: string
        cadastralReference:
          maxLength: 100
          minLength: 0
          type: string
        legalDescription:
          maxLength: 2000
          minLength: 0
          type: string
        zoningClassification:
          maxLength: 100
          minLength: 0
          type: string
        totalRooms:
          type: integer
          format: int32
        totalBedrooms:
          type: integer
          format: int32
        totalBathrooms:
          type: integer
          format: int32
        hasParking:
          type: boolean
        parkingSpaces:
          type: integer
          format: int32
        hasStorage:
          type: boolean
        storageArea:
          type: number
        hasElevator:
          type: boolean
        floorNumber:
          type: integer
          format: int32
        totalFloors:
          type: integer
          format: int32
        energyRating:
          maxLength: 10
          minLength: 0
          type: string
        heatingType:
          maxLength: 50
          minLength: 0
          type: string
        coolingType:
          maxLength: 50
          minLength: 0
          type: string
        restrictions:
          maxLength: 2000
          minLength: 0
          type: string
        amenities:
          type: string
        nearbyFacilities:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
...
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: The direction of sorting, either ASC or DESC.
          example: DESC
          default: DESC
      description: Represents a pagination request for retrieving paginated results, including page number, size, sort field, and direction.
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RangeObject'
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
    RangeObject:
      type: object
      properties:
        from:
          type: object
          description: Start value of the range
        to:
          type: object
          description: End value of the range
      description: Represents a range with from and to values
...
    FilterRequestGuaranteeValuationDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/GuaranteeValuationDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Generic filter request that includes both filter criteria and pagination
    PaginationResponseGuaranteeValuationDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/GuaranteeValuationDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...
    FilterRequestRealEstatePropertyDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/RealEstatePropertyDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Generic filter request that includes both filter criteria and pagination
    PaginationResponseRealEstatePropertyDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/RealEstatePropertyDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...